Add bike sharing example (#1)
This Broadway example aims to show a data processing pipeline
of events from a RabbitMQ queue. It simulates a bike sharing system
that is emitting coordinates of bikes moving through the city.

The sample was extracted from https://archive.ics.uci.edu/ml/datasets/GPS+Trajectories
and modified to match the bicycles ids.